extracts redirect behavior into its own file
